+++ packages/app/src/components/onboarding/index.ts
@@ -1,4 +1,5 @@
 export { PageIndicator } from './components/PageIndicator';
 export { OnboardingOrganize } from './OnboardingOrganize';
+export { OnboardingPage } from './OnboardingPage';
 export { OnboardingTeams } from './OnboardingTeams';
 export { OnboardingShare } from './OnboardingShare';
